remove incorrect cast

+++ tinyxml2.cpp
@@ -2635,7 +2635,7 @@ void XMLPrinter::Write( const char* data, size_t size )
         fwrite ( data , sizeof(char), size, _fp);
     }
     else {
-        char* p = _buffer.PushArr( static_cast<int>(size) ) - 1;   // back up over the null terminator.
+        char* p = _buffer.PushArr( size ) - 1;   // back up over the null terminator.
         memcpy( p, data, size );
         p[size] = 0;
     }
